+++ src/components/deprecated-tool-tip/__tests__/tool-tip.test.tsx
@@ -1,12 +1,12 @@
 import { render } from '@testing-library/react'
-import { ToolTip } from '../tool-tip'
+import { DeprecatedToolTip } from '../tool-tip'
 
-describe('ToolTip', () => {
+describe('DeprecatedToolTip', () => {
   it('should match a snapshot and render children', () => {
     const wrapper = render(
-      <ToolTip tip="hello there" defaultActive>
+      <DeprecatedToolTip tip="hello there" defaultActive>
         Hover here
-      </ToolTip>,
+      </DeprecatedToolTip>,
     )
     expect(wrapper.asFragment()).toMatchSnapshot()
   })
