+++ src/components/deprecated-tool-tip/index.ts
