+++ src/components/deprecated-tool-tip/tooltip.mdx
@@ -1,5 +1,4 @@
 import { Story, Canvas, Controls, Meta } from '@storybook/blocks'
-import { ToolTip } from './index'
 import { RenderHtmlMarkup } from '../../storybook/render-html-markup'
 import * as TooltipStories from './tooltip.stories'
