+++ src/components/deprecated-tool-tip/tooltip.stories.tsx
@@ -1,0 +1,18 @@
+import { DeprecatedToolTip } from './index'
+
+export default {
+  title: 'DeprecatedToolTip',
+  component: DeprecatedToolTip,
+}
+
+export const BasicUsage = {
+  render: ({}) => <DeprecatedToolTip tip="Some Data">Hover here</DeprecatedToolTip>,
+}
+
+export const DefaultActive = {
+  render: ({}) => (
+    <DeprecatedToolTip tip="Some Data" defaultActive>
+      Hover here
+    </DeprecatedToolTip>
+  ),
+}
