+++ src/index.ts
@@ -38,7 +38,7 @@ export * from './components/textarea'
 export * from './components/toggle'
 export * from './components/typography'
 export * from './components/form-layout'
-export * from './components/tool-tip'
+export * from './components/deprecated-tool-tip'
 export * from './components/input-error'
 export * from './components/file-input'
 export * from './components/placeholder-image'
